requirements.txt
python-telegram-bot==11.1.0
pymongo==3.7.1
psycopg2-binary